Moved client to the root folder
